api_server/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.6.1
Flask-Login
marshmallow==3.21.1
python-dotenv==1.0.1
PyMySQL==1.1.1
alembic==1.18.4
requests==2.32.5
google-genai==1.69.0
google-api-core==2.30.1
google-generativeai
python-socketio==5.16.1
python-engineio==4.13.1
PyYAML==6.0.3
Jinja2==3.1.6
Werkzeug==3.1.6